Refactor sensor.py into domain-specific modules
- Split 1250-line sensor.py into sensors/ directory with 7 modules
- sensor.py imports/re-exports from sensors/ (keeps async_setup_entry)
- Comprehensive tests for all sensor classes
- All tests pass (proves refactoring works)

Fixes #644
