+++ custom_components/localshift/sensor.py
@@ -1,31 +1,105 @@
 """Sensor platform for LocalShift integration."""
 
-from __future__ import annotations
-
-from typing import Any
-
-from homeassistant.components.sensor import (
-    SensorEntity,
-    SensorStateClass,
+from .sensors import (
+    AutomationReadySensor,
+    CheapChargeStopPriceSensor,
+    DecisionLagSensor,
+    DecisionLogSensor,
+    DecisionQualitySensor,
+    EffectiveCheapPriceSensor,
+    EntityHealthSensor,
+    ExcessSolarSensor,
+    ExtendedForecastAccuracySensor,
+    ForecastAccuracySensor,
+    ForecastDiagnosticsSensor,
+    ForecastHistorySensor,
+    ForecastPricesSensor,
+    ForecastStatusSensor,
+    IntegrationStatusSensor,
+    LearningDecisionHistorySensor,
+    LearningStatusSensor,
+    LoadShiftSignalSensor,
+    MinimumTargetSOCSensor,
+    NetElectricityCostSensor,
+    OptimizerPlanDetailedSensor,
+    OptimizerPlanGridSensor,
+    OptimizerPlanSensor,
+    OptimizerSummarySensor,
+    SolarBatteryForecastSensor,
+    SolarForecastAccuracySensor,
+    SolarWeightedAvgFITSensor,
 )
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from .sensors.base import LocalShiftSensorBase
 
-from .const import CONF_MINIMUM_TARGET_SOC, DEFAULT_MINIMUM_TARGET_SOC, DOMAIN
-from .coordinator import LocalShiftCoordinator
+# For backward compatibility - these were previously defined inline
+__all__ = [
+    "LocalShiftSensorBase",
+    "EffectiveCheapPriceSensor",
+    "CheapChargeStopPriceSensor",
+    "SolarWeightedAvgFITSensor",
+    "SolarBatteryForecastSensor",
+    "NetElectricityCostSensor",
+    "DecisionLogSensor",
+    "ForecastHistorySensor",
+    "OptimizerPlanSensor",
+    "ForecastPricesSensor",
+    "OptimizerPlanGridSensor",
+    "ForecastDiagnosticsSensor",
+    "MinimumTargetSOCSensor",
+    "ExcessSolarSensor",
+    "LoadShiftSignalSensor",
+    "ForecastAccuracySensor",
+    "IntegrationStatusSensor",
+    "EntityHealthSensor",
+    "LearningStatusSensor",
+    "DecisionQualitySensor",
+    "LearningDecisionHistorySensor",
+    "DecisionLagSensor",
+    "ExtendedForecastAccuracySensor",
+    "ForecastStatusSensor",
+    "AutomationReadySensor",
+    "OptimizerPlanDetailedSensor",
+    "OptimizerSummarySensor",
+    "SolarForecastAccuracySensor",
+]
 
 
-async def async_setup_entry(
-    hass: HomeAssistant,
-    entry: ConfigEntry,
-    async_add_entities: AddEntitiesCallback,
-) -> None:
+async def async_setup_entry(hass, entry, async_add_entities):
     """Set up LocalShift sensor entities."""
-    coordinator: LocalShiftCoordinator = entry.runtime_data
+    # Import here to avoid circular imports
+    from .sensors import (
+        AutomationReadySensor,
+        CheapChargeStopPriceSensor,
+        DecisionLagSensor,
+        DecisionLogSensor,
+        DecisionQualitySensor,
+        EffectiveCheapPriceSensor,
+        EntityHealthSensor,
+        ExcessSolarSensor,
+        ExtendedForecastAccuracySensor,
+        ForecastAccuracySensor,
+        ForecastDiagnosticsSensor,
+        ForecastHistorySensor,
+        ForecastPricesSensor,
+        ForecastStatusSensor,
+        IntegrationStatusSensor,
+        LearningDecisionHistorySensor,
+        LearningStatusSensor,
+        LoadShiftSignalSensor,
+        MinimumTargetSOCSensor,
+        NetElectricityCostSensor,
+        OptimizerPlanDetailedSensor,
+        OptimizerPlanGridSensor,
+        OptimizerPlanSensor,
+        OptimizerSummarySensor,
+        SolarBatteryForecastSensor,
+        SolarForecastAccuracySensor,
+        SolarWeightedAvgFITSensor,
+    )
 
-    entities: list[SensorEntity] = [
+    coordinator = entry.runtime_data
+
+    entities = [
         EffectiveCheapPriceSensor(coordinator, entry),
         CheapChargeStopPriceSensor(coordinator, entry),
         SolarWeightedAvgFITSensor(coordinator, entry),
@@ -70,1181 +144,3 @@ async def async_setup_entry(
     ]
 
     async_add_entities(entities)
-
-
-class LocalShiftSensorBase(SensorEntity):
-    """Base class for LocalShift sensors."""
-
-    _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        coordinator: LocalShiftCoordinator,
-        entry: ConfigEntry,
-    ) -> None:
-        """Initialise sensor."""
-        self.coordinator = coordinator
-        self._entry = entry
-        self._unsub: Any = None
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device information to link all entities under one device."""
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._entry.entry_id)},
-            name="LocalShift",
-            manufacturer="Custom",
-            model="Solar Battery Automation",
-            sw_version="0.0.2",
-        )
-
-    async def async_added_to_hass(self) -> None:
-        """Subscribe to coordinator updates."""
-        self._unsub = self.coordinator.async_add_listener(
-            self._handle_coordinator_update
-        )
-        # Set initial value
-        self._update_from_coordinator()
-
-    async def async_will_remove_from_hass(self) -> None:
-        """Unsubscribe from coordinator updates."""
-        if self._unsub:
-            self._unsub()
-
-    @callback
-    def _handle_coordinator_update(self) -> None:
-        """Handle updated data from coordinator."""
-        self._update_from_coordinator()
-        self.async_write_ha_state()
-
-    def _update_from_coordinator(self) -> None:
-        """Pull latest values from coordinator.data. Override in subclasses."""
-
-
-# ---------------------------------------------------------------------------
-# Sensor implementations
-# ---------------------------------------------------------------------------
-
-
-class EffectiveCheapPriceSensor(LocalShiftSensorBase):
-    """Dynamic cheap price threshold (urgency-adjusted)."""
-
-    _attr_unique_id = "localshift_price_cheap_effective"
-    _attr_name = "Price Cheap Effective"
-    _attr_icon = "mdi:tag-outline"
-    _attr_native_unit_of_measurement = "$/kWh"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        self._attr_native_value = round(self.coordinator.data.effective_cheap_price, 4)
-
-
-class CheapChargeStopPriceSensor(LocalShiftSensorBase):
-    """Effective threshold + deadband."""
-
-    _attr_unique_id = "localshift_price_cheap_charge_stop"
-    _attr_name = "Price Cheap Charge Stop"
-    _attr_icon = "mdi:tag-arrow-up-outline"
-    _attr_native_unit_of_measurement = "$/kWh"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        self._attr_native_value = round(
-            self.coordinator.data.cheap_charge_stop_price, 4
-        )
-
-
-class SolarWeightedAvgFITSensor(LocalShiftSensorBase):
-    """Solar-production-weighted average feed-in tariff."""
-
-    _attr_unique_id = "localshift_solar_weighted_avg_fit"
-    _attr_name = "Solar Weighted Avg FIT"
-    _attr_icon = "mdi:solar-power-variant"
-    _attr_native_unit_of_measurement = "$/kWh"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        self._attr_native_value = round(self.coordinator.data.solar_weighted_avg_fit, 4)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return additional attributes."""
-        return {
-            "total_solar_remaining_kwh": round(
-                self.coordinator.data.solar_remaining_kwh, 2
-            ),
-        }
-
-
-class SolarBatteryForecastSensor(LocalShiftSensorBase):
-    """Solar battery SOC forecast with detailed attributes."""
-
-    _attr_unique_id = "localshift_forecast_battery"
-    _attr_name = "Forecast Battery"
-    _attr_icon = "mdi:chart-line"
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        forecast = self.coordinator.data.solar_battery_forecast
-        self._attr_native_value = forecast.get("predicted_soc", 0)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return full forecast as attributes."""
-        return self.coordinator.data.solar_battery_forecast
-
-
-class NetElectricityCostSensor(LocalShiftSensorBase):
-    """Net electricity cost today (import cost - export revenue)."""
-
-    _attr_unique_id = "localshift_cost_electricity_net"
-    _attr_name = "Cost Electricity Net"
-    _attr_icon = "mdi:cash-register"
-    _attr_native_unit_of_measurement = "$"
-    _attr_state_class = SensorStateClass.TOTAL
-
-    def _update_from_coordinator(self) -> None:
-        d = self.coordinator.data
-        self._attr_native_value = round(d.grid_import_cost - d.grid_export_revenue, 2)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return cost breakdown."""
-        d = self.coordinator.data
-        return {
-            "grid_import_cost": round(d.grid_import_cost or 0.0, 2),
-            "grid_export_revenue": round(d.grid_export_revenue or 0.0, 2),
-            "battery_savings": round(d.battery_savings or 0.0, 2),
-            "battery_charge_cost": round(d.battery_charge_cost or 0.0, 2),
-        }
-
-
-class DecisionLogSensor(LocalShiftSensorBase):
-    """Battery mode change decision log."""
-
-    _attr_unique_id = "localshift_decision_log"
-    _attr_name = "Decision Log"
-    _attr_icon = "mdi:history"
-
-    def _update_from_coordinator(self) -> None:
-        log = self.coordinator.data.decision_log
-        if log:
-            self._attr_native_value = log[-1].get("reason", "")
-        else:
-            self._attr_native_value = "No decisions yet"
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return latest decision fields and recent history."""
-        log = self.coordinator.data.decision_log
-        attrs: dict[str, Any] = {"history": log[-10:]}
-        if log:
-            latest = log[-1]
-            attrs["reason"] = latest.get("reason", "")
-            attrs["soc"] = latest.get("soc")
-            attrs["buy_price"] = latest.get("buy_price")
-            attrs["sell_price"] = latest.get("sell_price")
-            attrs["timestamp"] = latest.get("timestamp")
-        return attrs
-
-
-class ForecastHistorySensor(LocalShiftSensorBase):
-    """Historical forecast predictions for planned vs actual comparison."""
-
-    _attr_unique_id = "localshift_forecast_history"
-    _attr_name = "Forecast History"
-    _attr_icon = "mdi:chart-line-variant"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with count of stored predictions."""
-        self._attr_native_value = len(self.coordinator.data.forecast_history)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return forecast history as attributes."""
-        return {"history": self.coordinator.data.forecast_history}
-
-
-class OptimizerPlanSensor(LocalShiftSensorBase):
-    """DP optimizer plan summary.
-
-    Shows the plan computed by the DP optimizer with slot actions and reason codes.
-    Replaces the legacy daily_forecast-based sensor (Phase 5, #447).
-    """
-
-    _attr_unique_id = "localshift_optimizer_plan"
-    _attr_name = "Optimizer Plan"
-    _attr_icon = "mdi:chart-bar"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with slot count."""
-        self._attr_native_value = len(self.coordinator.data.optimizer_decisions or [])
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return optimizer plan details."""
-        decisions = self.coordinator.data.optimizer_decisions or []
-        d = self.coordinator.data
-
-        slots = []
-        for dec in decisions:
-            slots.append({
-                "slot_idx": dec.get("slot_index"),
-                "action": dec.get("action"),
-                "reason_code": dec.get("reason_code"),
-                "objective_terms": dec.get("objective_terms", {}),
-            })
-
-        return {
-            "slots": slots,
-            "total_slots": len(slots),
-            "forecast_horizon_hours": d.forecast_horizon_hours,
-            "planner": "DP_OPTIMIZER",
-        }
-
-
-class ForecastPricesSensor(LocalShiftSensorBase):
-    """Price forecast data for history collection.
-
-    Split from DailyForecastSensor to stay under 16KB limit (Issue #37).
-    Provides buy/sell price time series for analysis and dashboards.
-    Phase 5 (#447): Reads from optimizer_decisions with fallback to raw forecasts.
-    """
-
-    _attr_unique_id = "localshift_forecast_prices"
-    _attr_name = "Forecast Prices"
-    _attr_icon = "mdi:currency-usd"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with current effective cheap price."""
-        self._attr_native_value = round(self.coordinator.data.effective_cheap_price, 4)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return price forecast data."""
-        d = self.coordinator.data
-        decisions = d.optimizer_decisions or []
-
-        buy_prices = []
-        sell_prices = []
-
-        if decisions:
-            for dec in decisions:
-                ts = dec.get("timestamp_iso", "")
-                time_str = ts[11:16] if len(ts) >= 16 else ""
-                buy_prices.append({
-                    "time": time_str,
-                    "price": dec.get("buy_price"),
-                })
-                sell_prices.append({
-                    "time": time_str,
-                    "price": dec.get("sell_price"),
-                })
-        else:
-            for slot in d.general_forecast:
-                ts = slot.get("timestamp", "")
-                time_str = ts[11:16] if len(ts) >= 16 else ""
-                buy_prices.append({
-                    "time": time_str,
-                    "price": slot.get("price"),
-                })
-            for slot in d.feed_in_forecast:
-                ts = slot.get("timestamp", "")
-                time_str = ts[11:16] if len(ts) >= 16 else ""
-                sell_prices.append({
-                    "time": time_str,
-                    "price": slot.get("price"),
-                })
-
-        return {
-            "buy_prices": buy_prices,
-            "sell_prices": sell_prices,
-            "effective_cheap_price": round(d.effective_cheap_price, 4),
-            "cheap_charge_stop_price": round(d.cheap_charge_stop_price, 4),
-            "forecast_import_cost": round(d.forecast_import_cost or 0.0, 2),
-            "forecast_export_revenue": round(d.forecast_export_revenue or 0.0, 2),
-            "forecast_net_cost": round(d.forecast_net_cost or 0.0, 2),
-            "forecast_grid_charge_cost": round(d.forecast_grid_charge_cost or 0.0, 2),
-            "forecast_proactive_export_revenue": round(
-                d.forecast_proactive_export_revenue or 0.0, 2
-            ),
-        }
-
-
-class OptimizerPlanGridSensor(LocalShiftSensorBase):
-    """Grid interaction projections from DP optimizer.
-
-    Shows projected import/export from the optimizer plan.
-    Replaces the legacy forecast_grid sensor (Phase 5, #447).
-    """
-
-    _attr_unique_id = "localshift_optimizer_plan_grid"
-    _attr_name = "Optimizer Plan Grid"
-    _attr_icon = "mdi:transmission-tower"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with projected net cost."""
-        summary = self.coordinator.data.optimizer_summary or {}
-        self._attr_native_value = round(summary.get("projected_net_cost", 0.0), 3)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return grid interaction projections."""
-        d = self.coordinator.data
-        decisions = d.optimizer_decisions or []
-        summary = d.optimizer_summary or {}
-
-        projected_import = summary.get("projected_import_kwh", 0.0)
-        projected_export = summary.get("projected_export_kwh", 0.0)
-        projected_net_cost = summary.get("projected_net_cost", 0.0)
-
-        action_counts: dict[str, int] = {}
-        for dec in decisions:
-            action = dec.get("action", "UNKNOWN")
-            action_counts[action] = action_counts.get(action, 0) + 1
-
-        return {
-            "projected_import_kwh": round(projected_import, 3),
-            "projected_export_kwh": round(projected_export, 3),
-            "projected_net_cost": round(projected_net_cost, 3),
-            "action_breakdown": action_counts,
-            "planner": "DP_OPTIMIZER",
-        }
-
-
-class ForecastDiagnosticsSensor(LocalShiftSensorBase):
-    """Diagnostic and debug data for the forecast system.
-
-    Split from DailyForecastSensor to stay under 16KB limit (Issue #37).
-    Contains consumption profiles, weather correlation, and debug fields.
-    """
-
-    _attr_unique_id = "localshift_forecast_diagnostics"
-    _attr_name = "Forecast Diagnostics"
-    _attr_icon = "mdi:bug-outline"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with consumption source."""
-        self._attr_native_value = self.coordinator.data.consumption_source
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return diagnostic and debug data."""
-        sample_counts = self.coordinator.data.consumption_hourly_sample_counts
-        profile_kw = self.coordinator.data.consumption_hourly_profile_kw
-        source_counts = self.coordinator.data.forecast_consumption_source_counts
-
-        return {
-            # Consumption profile information
-            "consumption_source": self.coordinator.data.consumption_source,
-            "consumption_statistic_id": self.coordinator.data.consumption_statistic_id,
-            "consumption_profile_hours": self.coordinator.data.consumption_profile_hours,
-            "consumption_fallback_hours": self.coordinator.data.consumption_fallback_hours,
-            "forecast_consumption_source_counts": dict(source_counts),
-            "consumption_hourly_sample_counts": {
-                str(hour): count for hour, count in sorted(sample_counts.items())
-            },
-            "consumption_hourly_profile_kw": {
-                str(hour): value for hour, value in sorted(profile_kw.items())
-            },
-            # Day-of-week aware consumption profiles (issue-60)
-            "consumption_profile_type": self.coordinator.data.consumption_profile_type,
-            "forecast_profile_selected": self.coordinator.data.forecast_profile_selected,
-            "weekday_sample_counts": {
-                str(hour): count
-                for hour, count in sorted(
-                    self.coordinator.data.weekday_sample_counts.items()
-                )
-            },
-            "weekend_sample_counts": {
-                str(hour): count
-                for hour, count in sorted(
-                    self.coordinator.data.weekend_sample_counts.items()
-                )
-            },
-            "weekday_hourly_profile_kw": {
-                str(hour): value
-                for hour, value in sorted(
-                    self.coordinator.data.weekday_hourly_profile_kw.items()
-                )
-            },
-            "weekend_hourly_profile_kw": {
-                str(hour): value
-                for hour, value in sorted(
-                    self.coordinator.data.weekend_hourly_profile_kw.items()
-                )
-            },
-            # Recent load data
-            "recent_load_1hr_kw": round(self.coordinator.data.recent_load_1hr_kw, 3),
-            "recent_load_1hr_statistic_id": self.coordinator.data.recent_load_1hr_statistic_id,
-            "recent_load_1hr_samples": self.coordinator.data.recent_load_1hr_samples,
-            "recent_load_1hr_last_error": self.coordinator.data.recent_load_1hr_last_error,
-            "current_load_kw": round(self.coordinator.data.load_power_kw, 3),
-            # Debug fields for troubleshooting mode decisions
-            "debug_forecast_slot_found": self.coordinator.data.debug_forecast_slot_found,
-            "debug_forecast_slot_time": self.coordinator.data.debug_forecast_slot_time,
-            "debug_first_forecast_slot_time": self.coordinator.data.debug_first_forecast_slot_time,
-            "debug_time_gap_seconds": round(
-                self.coordinator.data.debug_time_gap_seconds, 1
-            ),
-            "debug_mode_source": self.coordinator.data.debug_mode_source,
-            # Export permission
-            "allow_export": self.coordinator.data.allow_export,
-            # Weather correlation visibility (Issue #61)
-            "weather_entity_id": self.coordinator.data.weather_entity_id,
-            "weather_temperature_current": self.coordinator.data.weather_temperature_current,
-            "weather_temperature_forecast": {
-                str(hour): temp
-                for hour, temp in sorted(
-                    self.coordinator.data.weather_temperature_forecast.items()
-                )
-            },
-            "weather_condition": self.coordinator.data.weather_condition,
-            "weather_correlation_confidence": self.coordinator.data.weather_correlation_confidence,
-            "weather_adjustment_applied": self.coordinator.data.weather_adjustment_applied,
-            "weather_learning_enabled": self.coordinator.data.weather_learning_enabled,
-            "weather_cooling_coefficient": round(
-                self.coordinator.data.weather_cooling_coefficient, 4
-            ),
-            "weather_heating_coefficient": round(
-                self.coordinator.data.weather_heating_coefficient, 4
-            ),
-            "weather_sample_count": self.coordinator.data.weather_sample_count,
-            # Issue #500: Load forecast slots runtime data
-            "load_forecast_slots_sample": {
-                "count": len(self.coordinator.data.load_forecast_slots),
-                "first_5": [
-                    round(v, 3) for v in self.coordinator.data.load_forecast_slots[:5]
-                ]
-                if self.coordinator.data.load_forecast_slots
-                else [],
-                "noon_idx_4": round(self.coordinator.data.load_forecast_slots[4], 3)
-                if len(self.coordinator.data.load_forecast_slots) > 4
-                else None,
-                "noon_idx_5": round(self.coordinator.data.load_forecast_slots[5], 3)
-                if len(self.coordinator.data.load_forecast_slots) > 5
-                else None,
-                "idx_8": round(self.coordinator.data.load_forecast_slots[8], 3)
-                if len(self.coordinator.data.load_forecast_slots) > 8
-                else None,
-            },
-            # Issue #500: Adaptive params values
-            "adaptive_params_values": dict(self.coordinator.data.adaptive_params.values)
-            if self.coordinator.data.adaptive_params
-            else {},
-        }
-
-
-class MinimumTargetSOCSensor(LocalShiftSensorBase):
-    """Minimum target SOC for discharge modes (base reserve)."""
-
-    _attr_unique_id = "localshift_target_soc_minimum"
-    _attr_name = "Target SOC Minimum"
-    _attr_icon = "mdi:battery-charging-20"
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Read minimum target SOC from config options."""
-        self._attr_native_value = float(
-            self._entry.options.get(CONF_MINIMUM_TARGET_SOC, DEFAULT_MINIMUM_TARGET_SOC)
-        )
-
-
-# ---------------------------------------------------------------------------
-# Excess Solar Load Shifting Sensors (backlog-high-017)
-# ---------------------------------------------------------------------------
-
-
-class ExcessSolarSensor(LocalShiftSensorBase):
-    """Forecasted excess solar energy available for discretionary loads."""
-
-    _attr_unique_id = "localshift_excess_solar_kwh"
-    _attr_name = "Excess Solar"
-    _attr_icon = "mdi:solar-power"
-    _attr_native_unit_of_measurement = "kWh"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with total excess until battery full."""
-        self._attr_native_value = round(
-            self.coordinator.data.excess_until_battery_full_kwh, 2
-        )
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed excess solar attributes."""
-        d = self.coordinator.data
-        return {
-            "excess_current_hour_kwh": round(d.excess_solar_current_hour_kwh, 2),
-            "excess_next_2h_kwh": round(d.excess_solar_next_2h_kwh, 2),
-            "excess_next_4h_kwh": round(d.excess_solar_next_4h_kwh, 2),
-            "excess_until_battery_full_kwh": round(d.excess_until_battery_full_kwh, 2),
-            "excess_until_negative_fit_kwh": round(d.excess_until_negative_fit_kwh, 2),
-            "time_until_battery_full_minutes": d.time_until_battery_full_minutes,
-            "negative_fit_window_start": (
-                d.negative_fit_window_start.isoformat()
-                if d.negative_fit_window_start
-                else None
-            ),
-            "negative_fit_window_duration_minutes": d.negative_fit_window_duration_minutes,
-            "can_add_load_now": d.can_add_load_now,
-            "safe_additional_load_kw": round(d.safe_additional_load_kw, 1),
-            "forecast_confidence": d.load_shift_confidence,
-            "current_excess_rate_kw": round(d.current_excess_rate_kw, 2),
-        }
-
-
-class LoadShiftSignalSensor(LocalShiftSensorBase):
-    """Actionable signal for automations indicating what to do with discretionary loads."""
-
-    _attr_unique_id = "localshift_load_shift_signal"
-    _attr_name = "Load Shift Signal"
-    _attr_icon = "mdi:transfer"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with current load shift signal."""
-        self._attr_native_value = self.coordinator.data.load_shift_signal
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return signal details."""
-        d = self.coordinator.data
-        return {
-            "recommended_additional_kw": round(d.load_shift_recommended_kw, 1),
-            "recommended_duration_minutes": d.load_shift_recommended_duration_minutes,
-            "signal_reason": d.load_shift_reason,
-            "signal_confidence": d.load_shift_confidence,
-            "current_excess_rate_kw": round(d.current_excess_rate_kw, 2),
-            "grid_charge_risk": d.grid_charge_risk,
-            "safe_additional_load_kw": round(d.safe_additional_load_kw, 1),
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on signal state."""
-        signal = self._attr_native_value
-        if signal == "INCREASE_LOAD":
-            return "mdi:arrow-up-bold"
-        elif signal == "REDUCE_LOAD":
-            return "mdi:arrow-down-bold"
-        elif signal == "MAINTAIN_LOAD":
-            return "mdi:check-circle"
-        else:  # HOLD
-            return "mdi:pause-circle"
-
-
-# ---------------------------------------------------------------------------
-# Forecast Accuracy Sensor (Issue #37 Phase 2)
-# ---------------------------------------------------------------------------
-
-
-class ForecastAccuracySensor(LocalShiftSensorBase):
-    """Compares past forecast predictions with actual outcomes.
-
-    Reads from HA history (which now works thanks to Phase 1) and computes
-    accuracy metrics for SOC, price, and other predictions.
-    """
-
-    _attr_unique_id = "localshift_forecast_accuracy"
-    _attr_name = "Forecast Accuracy"
-    _attr_icon = "mdi:target"
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with overall accuracy percentage."""
-        accuracy = self.coordinator.data.forecast_accuracy_soc_1h
-        self._attr_native_value = round(accuracy, 1) if accuracy is not None else None
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed accuracy metrics."""
-        d = self.coordinator.data
-        return {
-            # SOC prediction errors (predicted - actual)
-            "soc_error_15min": round(d.forecast_error_soc_15min, 1),
-            "soc_error_1h": round(d.forecast_error_soc_1h, 1),
-            "soc_error_4h": round(d.forecast_error_soc_4h, 1),
-            # SOC accuracy percentages (100 - abs error, clamped)
-            "soc_accuracy_15min": round(d.forecast_accuracy_soc_15min, 1)
-            if d.forecast_accuracy_soc_15min is not None
-            else None,
-            "soc_accuracy_1h": round(d.forecast_accuracy_soc_1h, 1)
-            if d.forecast_accuracy_soc_1h is not None
-            else None,
-            "soc_accuracy_4h": round(d.forecast_accuracy_soc_4h, 1)
-            if d.forecast_accuracy_soc_4h is not None
-            else None,
-            # Price prediction errors
-            "buy_price_error_1h": round(d.forecast_error_buy_price_1h, 4),
-            "sell_price_error_1h": round(d.forecast_error_sell_price_1h, 4),
-            # Comparison metadata
-            "comparisons_made": d.forecast_comparisons_made,
-            "last_comparison_time": d.forecast_last_comparison_time,
-            # Forecast history persistence (Issue #131)
-            "first_prediction_time": d.forecast_first_prediction_time,
-            "history_count": d.forecast_history_count,
-        }
-
-
-# ---------------------------------------------------------------------------
-# Entity Health and Error Tracking Sensors (Issue #94)
-# ---------------------------------------------------------------------------
-
-
-class IntegrationStatusSensor(LocalShiftSensorBase):
-    """Overall integration health status.
-
-    Shows "ok", "degraded", or "error" based on entity availability.
-    Provides a simple status indicator for dashboards.
-    """
-
-    _attr_unique_id = "localshift_integration_status"
-    _attr_name = "Integration Status"
-    _attr_icon = "mdi:check-circle"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with current integration status."""
-        self._attr_native_value = self.coordinator.data.integration_status
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return status details."""
-        d = self.coordinator.data
-        return {
-            "message": d.integration_status_message,
-            "error_count": len(d.entity_errors),
-            "warning_count": len(d.entity_warnings),
-            "required_entities_healthy": d.required_entities_healthy,
-            "errors": d.entity_errors,
-            "warnings": d.entity_warnings,
-            "last_check": d.last_entity_check,
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on status."""
-        status = self._attr_native_value
-        if status == "ok":
-            return "mdi:check-circle"
-        elif status == "degraded":
-            return "mdi:alert-circle"
-        else:  # error
-            return "mdi:close-circle"
-
-
-class EntityHealthSensor(LocalShiftSensorBase):
-    """Detailed health status for all tracked entities.
-
-    Provides per-entity health information as JSON attributes
-    for detailed diagnostics and troubleshooting.
-    """
-
-    _attr_unique_id = "localshift_entity_health"
-    _attr_name = "Entity Health"
-    _attr_icon = "mdi:heart-pulse"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with count of healthy entities."""
-        dep_health = self.coordinator.data.entity_health
-        ls_health = self.coordinator.data.localshift_entity_health
-        all_entities = {**dep_health, **ls_health}
-        healthy_count = sum(1 for e in all_entities.values() if e.get("status") == "ok")
-        total_count = len(all_entities)
-        self._attr_native_value = f"{healthy_count}/{total_count}"
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed entity health."""
-        dep_health = self.coordinator.data.entity_health
-        ls_health = self.coordinator.data.localshift_entity_health
-        return {
-            "entities": dep_health,  # legacy key for backward compatibility
-            "dependencies": dep_health,
-            "localshift_entities": ls_health,
-            "errors": self.coordinator.data.entity_errors,
-            "warnings": self.coordinator.data.entity_warnings,
-            "summary": {
-                "dependencies": {
-                    "total": len(dep_health),
-                    "healthy": sum(
-                        1 for e in dep_health.values() if e.get("status") == "ok"
-                    ),
-                },
-                "localshift": {
-                    "total": len(ls_health),
-                    "healthy": sum(
-                        1 for e in ls_health.values() if e.get("status") == "ok"
-                    ),
-                    "by_category": {
-                        "required": sum(
-                            1
-                            for e in ls_health.values()
-                            if e.get("category") == "required"
-                            and e.get("status") == "ok"
-                        ),
-                        "optional": sum(
-                            1
-                            for e in ls_health.values()
-                            if e.get("category") == "optional"
-                            and e.get("status") == "ok"
-                        ),
-                    },
-                },
-            },
-        }
-
-
-# ---------------------------------------------------------------------------
-# Learning System Sensors (Issue #170 Phase 5)
-# ---------------------------------------------------------------------------
-
-
-class LearningStatusSensor(LocalShiftSensorBase):
-    """Current learning system status and parameter values.
-
-    Shows the learning phase: 'observing', 'tuning', 'optimizing', or 'disabled'.
-    Provides visibility into adaptive parameters and their confidence scores.
-    """
-
-    _attr_unique_id = "localshift_learning_status"
-    _attr_name = "Learning Status"
-    _attr_icon = "mdi:brain"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with current learning status."""
-        self._attr_native_value = self.coordinator.data.learning_status
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return learning system details."""
-        d = self.coordinator.data
-        metrics = d.performance_metrics
-        return {
-            "total_decisions_today": metrics.total_decisions_today,
-            "avg_decision_score_today": round(metrics.avg_decision_score_today, 3),
-            "avg_decision_score_7d": round(metrics.avg_decision_score_7d, 3),
-            "cost_trend": metrics.cost_trend,
-            "mode_durations_today": metrics.mode_durations_today,
-            "mode_cost_attribution": {
-                k: round(v, 3) for k, v in metrics.mode_cost_attribution.items()
-            },
-            # Optimization weights for multi-objective scoring (Issue #170 Phase 4)
-            "optimization_weights": d.optimization_weights,
-            # Active contextual adjustments (Issue #170 Phase 4)
-            "contextual_adjustments_active": d.contextual_adjustments_active,
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on status."""
-        status = self._attr_native_value
-        if status == "optimizing":
-            return "mdi:brain"
-        elif status == "tuning":
-            return "mdi:tune"
-        elif status == "observing":
-            return "mdi:eye"
-        else:  # disabled
-            return "mdi:brain-off"
-
-
-class DecisionQualitySensor(LocalShiftSensorBase):
-    """Rolling decision quality score.
-
-    Shows today's average decision quality as a percentage.
-    Provides detailed metrics about decision outcomes.
-    """
-
-    _attr_unique_id = "localshift_decision_quality"
-    _attr_name = "Decision Quality"
-    _attr_icon = "mdi:chart-line"
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with today's average decision quality."""
-        score = self.coordinator.data.performance_metrics.avg_decision_score_today
-        self._attr_native_value = round(score * 100, 1)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed quality metrics."""
-        metrics = self.coordinator.data.performance_metrics
-        return {
-            "total_decisions_today": metrics.total_decisions_today,
-            "avg_score_7d": round(metrics.avg_decision_score_7d * 100, 1),
-            "cost_trend": metrics.cost_trend,
-            "grid_charge_efficiency": round(metrics.grid_charge_efficiency * 100, 1),
-            "export_loss_ratio": round(metrics.export_loss_ratio * 100, 1),
-            "unnecessary_grid_charge_kwh": round(
-                metrics.unnecessary_grid_charge_kwh, 2
-            ),
-            "mode_durations_today": metrics.mode_durations_today,
-            "mode_cost_attribution": {
-                k: round(v, 3) for k, v in metrics.mode_cost_attribution.items()
-            },
-        }
-
-
-class LearningDecisionHistorySensor(LocalShiftSensorBase):
-    """Recent decision history with outcomes.
-
-    Shows count of decisions in last 24h and provides
-    the most recent decision records for debugging.
-    """
-
-    _attr_unique_id = "localshift_learning_decision_history"
-    _attr_name = "Learning Decision History"
-    _attr_icon = "mdi:history"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with count of recent decisions."""
-        self._attr_native_value = len(self.coordinator.data.recent_decision_log)
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return recent decision records."""
-        return {
-            "decisions": self.coordinator.data.recent_decision_log[-20:],
-        }
-
-
-# ---------------------------------------------------------------------------
-# Decision-to-Implementation Lag Sensor (Issue #501)
-# ---------------------------------------------------------------------------
-
-
-class DecisionLagSensor(LocalShiftSensorBase):
-    """Measures time from decision to validated implementation.
-
-    Shows the lag in seconds between when a battery mode decision is made
-    and when the hardware validates the change.
-    """
-
-    _attr_unique_id = "localshift_decision_lag"
-    _attr_name = "Decision Lag"
-    _attr_icon = "mdi:timer-outline"
-    _attr_native_unit_of_measurement = "s"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with current lag in seconds."""
-        lag = self.coordinator.data.decision_lag_seconds
-        if lag is not None:
-            self._attr_native_value = round(lag, 2)
-        else:
-            self._attr_native_value = None
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return lag history and statistics."""
-        d = self.coordinator.data
-        history = d.decision_lag_history or []
-
-        # Calculate statistics from history
-        if history:
-            lag_values = [h["lag_seconds"] for h in history]
-            avg_lag = sum(lag_values) / len(lag_values)
-            max_lag = max(lag_values)
-            min_lag = min(lag_values)
-        else:
-            avg_lag = None
-            max_lag = None
-            min_lag = None
-
-        return {
-            "current_lag": round(d.decision_lag_seconds, 2)
-            if d.decision_lag_seconds is not None
-            else None,
-            "last_transition": history[-1] if history else None,
-            "history": history[-20:],  # Last 20 entries
-            "avg_lag_24h": round(avg_lag, 2) if avg_lag is not None else None,
-            "max_lag_24h": round(max_lag, 2) if max_lag is not None else None,
-            "min_lag_24h": round(min_lag, 2) if min_lag is not None else None,
-            "total_transitions": len(history),
-            "decision_timestamp": d.decision_timestamp.isoformat()
-            if d.decision_timestamp
-            else None,
-            "implementation_timestamp": d.implementation_timestamp.isoformat()
-            if d.implementation_timestamp
-            else None,
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on lag status."""
-        if self.coordinator.data.decision_timestamp is not None:
-            return "mdi:timer-sand"  # Decision pending
-        return "mdi:timer-outline"  # Completed
-
-
-class ExtendedForecastAccuracySensor(LocalShiftSensorBase):
-    """Extended forecast accuracy with long-term metrics.
-
-    Issue #270: Multi-horizon validation with bias detection.
-    Shows 24h, 7d, and 30d accuracy metrics.
-    """
-
-    _attr_unique_id = "localshift_extended_forecast_accuracy"
-    _attr_name = "Extended Forecast Accuracy"
-    _attr_icon = "mdi:chart-timeline-variant"
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def _update_from_coordinator(self) -> None:
-        """Update with 24h accuracy."""
-        self._attr_native_value = round(
-            self.coordinator.data.extended_accuracy_metrics.accuracy_24h, 1
-        )
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return extended accuracy metrics."""
-        m = self.coordinator.data.extended_accuracy_metrics
-        return {
-            "accuracy_24h": round(m.accuracy_24h, 1),
-            "accuracy_7d": round(m.accuracy_7d, 1),
-            "accuracy_30d": round(m.accuracy_30d, 1),
-            "bias": round(m.bias, 2),
-            "mape": round(m.mape, 2),
-            "sample_count": m.sample_count,
-            "last_updated": m.last_updated.isoformat() if m.last_updated else None,
-        }
-
-
-# ---------------------------------------------------------------------------
-# Forecast Status Sensor (Issue #319)
-# ---------------------------------------------------------------------------
-
-
-class ForecastStatusSensor(LocalShiftSensorBase):
-    """Forecast data readiness status.
-
-    Issue #319: Shows whether Solcast forecast data is available and valid.
-    Prevents premature grid charging decisions when forecast data is not ready.
-
-    States:
-    - "ready": Solcast data is available and sufficient for forecasting
-    - "partial": Solcast data is available but incomplete (fewer than 8 entries)
-    - "stale": Solcast entity not found or state is unknown/unavailable
-    - "initializing": Initial state before first check
-    """
-
-    _attr_unique_id = "localshift_forecast_status"
-    _attr_name = "Forecast Status"
-    _attr_icon = "mdi:weather-sunny"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with forecast status."""
-        self._attr_native_value = self.coordinator.data.forecast_status
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return forecast status details."""
-        d = self.coordinator.data
-        return {
-            "forecast_ready": d.forecast_ready,
-            "solcast_today_entries": len(d.solcast_today),
-            "solcast_tomorrow_entries": len(d.solcast_tomorrow),
-            "debug_mode_source": d.debug_mode_source,
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on status."""
-        status = self._attr_native_value
-        if status == "ready":
-            return "mdi:check-circle"
-        elif status == "partial":
-            return "mdi:alert-circle"
-        elif status == "stale":
-            return "mdi:close-circle"
-        else:  # initializing
-            return "mdi:weather-sunny-alert"
-
-
-# ---------------------------------------------------------------------------
-# Automation Ready Sensor (Issue #349)
-# ---------------------------------------------------------------------------
-
-
-class AutomationReadySensor(LocalShiftSensorBase):
-    """Automation ready status.
-
-    Issue #349: Shows whether all required inputs are valid for automation.
-    Prevents incorrect mode inference at startup when entities haven't populated.
-
-    States:
-    - "ready": All required inputs are valid
-    - "not_ready": One or more inputs are missing/invalid
-    """
-
-    _attr_unique_id = "localshift_automation_ready"
-    _attr_name = "Automation Ready"
-    _attr_icon = "mdi:check-decagram"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with automation ready status."""
-        if self.coordinator.data.automation_ready:
-            self._attr_native_value = "ready"
-        else:
-            self._attr_native_value = "not_ready"
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed status."""
-        d = self.coordinator.data
-        return {
-            "automation_ready": d.automation_ready,
-            "status_checks": d.automation_ready_status,
-            "missing_inputs": d.automation_ready_missing,
-            "soc": d.soc,
-            "operation_mode": d.operation_mode,
-            "backup_reserve": d.backup_reserve,
-            "prices_available": d.prices_available,
-            "forecast_status": d.forecast_status,
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on status."""
-        if self._attr_native_value == "ready":
-            return "mdi:check-decagram"
-        else:
-            return "mdi:decagram-outline"
-
-
-# ---------------------------------------------------------------------------
-# Optimizer Shadow Telemetry Sensors (Issue #403 Phase 1)
-# ---------------------------------------------------------------------------
-
-
-class OptimizerPlanDetailedSensor(LocalShiftSensorBase):
-    """DP optimizer detailed plan with full per-slot data.
-
-    Shows the complete plan with all slot details for diagnostics.
-    Phase 5 (#447): Renamed from OptimizerShadowPlanSensor, reads optimizer_decisions.
-    """
-
-    _attr_unique_id = "localshift_optimizer_plan_detailed"
-    _attr_name = "Optimizer Plan Detailed"
-    _attr_icon = "mdi:format-list-bulleted"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with plan status."""
-        summary = self.coordinator.data.optimizer_summary or {}
-        if not summary or not summary.get("enabled", False):
-            self._attr_native_value = "disabled"
-        elif not summary.get("success", False):
-            self._attr_native_value = "error"
-        else:
-            self._attr_native_value = "computed"
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed plan data."""
-        d = self.coordinator.data
-        decisions = d.optimizer_decisions or []
-        summary = d.optimizer_summary or {}
-
-        return {
-            "enabled": summary.get("enabled", False),
-            "success": summary.get("success", False),
-            "error_message": summary.get("error_message"),
-            "decisions": decisions,
-            "total_slots": len(decisions),
-            "forecast_horizon_hours": d.forecast_horizon_hours,
-            "computed_at": summary.get("cycle_timestamp_iso")
-            or summary.get("computed_at"),
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on status."""
-        status = self._attr_native_value
-        if status == "computed":
-            return "mdi:format-list-bulleted"
-        elif status == "error":
-            return "mdi:alert-circle-outline"
-        else:
-            return "mdi:minus-circle-outline"
-
-
-class OptimizerSummarySensor(LocalShiftSensorBase):
-    """DP optimizer run summary.
-
-    Phase 5 (#447): Renamed from OptimizerShadowSummarySensor, reads optimizer_summary.
-    """
-
-    _attr_unique_id = "localshift_optimizer_summary"
-    _attr_name = "Optimizer Summary"
-    _attr_icon = "mdi:chart-box-outline"
-
-    def _update_from_coordinator(self) -> None:
-        """Update with optimizer run summary."""
-        summary = self.coordinator.data.optimizer_summary or {}
-        if not summary or not summary.get("enabled", False):
-            self._attr_native_value = "disabled"
-        elif summary.get("success", False):
-            self._attr_native_value = "success"
-        else:
-            self._attr_native_value = "failed"
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return optimizer summary details."""
-        summary = self.coordinator.data.optimizer_summary or {}
-        return {
-            "enabled": summary.get("enabled", False),
-            "success": summary.get("success", False),
-            "error_message": summary.get("error_message"),
-            "computed_at": summary.get("cycle_timestamp_iso")
-            or summary.get("computed_at"),
-            "config_options": summary.get("config_options", {}),
-            "parity_completeness_pct": summary.get("parity_completeness_pct"),
-            "parity_defaulted_fields": summary.get("parity_defaulted_fields", {}),
-            "alignment_valid": summary.get("alignment_valid"),
-            "alignment_issues": summary.get("alignment_issues", []),
-            "alignment_warnings": summary.get("alignment_warnings", []),
-            "planner_version": summary.get("planner_version"),
-            "cycle_id": summary.get("cycle_id"),
-            "solve_time_seconds": summary.get("solve_time_seconds"),
-            "projected_net_cost": summary.get("projected_net_cost"),
-            "terminal_shortfall_pct": summary.get("terminal_shortfall_pct"),
-            "initial_soc_pct": summary.get("initial_soc_pct"),
-        }
-
-    @property
-    def icon(self) -> str:
-        """Return icon based on status."""
-        status = self._attr_native_value
-        if status == "success":
-            return "mdi:check-circle-outline"
-        elif status == "failed":
-            return "mdi:alert-circle-outline"
-        else:
-            return "mdi:minus-circle-outline"
-
-
-class SolarForecastAccuracySensor(LocalShiftSensorBase):
-    """Solar forecast accuracy metrics for learning system."""
-
-    _attr_unique_id = "localshift_solar_forecast_accuracy"
-    _attr_name = "Solar Forecast Accuracy"
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-    _attr_icon = "mdi:solar-power-variant"
-
-    def _update_from_coordinator(self) -> None:
-        self._attr_native_value = getattr(
-            self.coordinator.data, "solar_forecast_accuracy", 100.0
-        )
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any] | None:
-        return getattr(self.coordinator.data, "solar_bias_metrics", {})
