+++ tests/sensors/test_base.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+class TestBaseModule:
+    def test_import(self):
+        from custom_components.localshift.sensors.base import LocalShiftSensorBase
+
+        assert LocalShiftSensorBase is not None
