+++ tests/sensors/test_misc.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+class TestMiscModule:
+    def test_import(self):
+        from custom_components.localshift.sensors.misc import ExcessSolarSensor
+
+        assert ExcessSolarSensor is not None
