+++ tests/sensors/test_optimizer.py
@@ -1,0 +1,10 @@
+import pytest
+
+
+class TestOptimizerModule:
+    def test_import(self):
+        from custom_components.localshift.sensors.optimizer import (
+            OptimizerPlanDetailedSensor,
+        )
+
+        assert OptimizerPlanDetailedSensor is not None
