+++ tests/sensors/test_pricing.py
@@ -1,0 +1,10 @@
+import pytest
+
+
+class TestPricingModule:
+    def test_import(self):
+        from custom_components.localshift.sensors.pricing import (
+            EffectiveCheapPriceSensor,
+        )
+
+        assert EffectiveCheapPriceSensor is not None
